Improve build and install
This is a huge commit, I know...
Turns out, the CMake config files were not correct, so I could not import this
project from another CMake project. I took a break and read "Professional CMake,
A practical guide" from Craig Scott. I have now a much better understanding of
the language and the build and install steps. I still need to look at cross-
compilation and relocatable installations regarding the runtime path aka RPATH.

* The project noa correctly exports its targets and encodes everything in the
  noaConfig.cmake as it is supposed to. This file saves the parameters that
  were used during the build for convenience, but the params are cache variable
  so they can be overwritten if desired. Some options cannot be changed between
  build and install, and this is usually encoded in the target names.

* CBLAS/LAPACKE search is much more flexible now. We don't require OpenBLAS.
  If OpenBLAS, our module FindCBLAS.cmake will detect it and try to figure out
  the threading model so that we can easily know what we use.

* FFTW3 with or without thread is actually important. If built with threads, the
  library will need the threaded version at runtime. So for the import, we make
  sure this is OK by having the name of the threading model in the target name.
  That way, we create a "hard" dependency.

* The dependencies that can be private are now private. In the future it would be
  great to have FFTW3 private, for licence purposes.

Tags: [fix][cmake]
Compile: yes
Tests: yes

+++ tests/noa/cpu/geometry/TestCPUTransformProject.cpp
@@ -46,7 +46,7 @@ namespace {
 
                     for (int64_t z = start; z < end; ++z) {
                         // Round coordinate to have the XYZ output index.
-                        const float3_t output_idx{math::floor(output_coords[0]) + z,
+                        const float3_t output_idx{math::floor(output_coords[0]) + static_cast<float>(z),
                                                   math::floor(output_coords[1]),
                                                   math::floor(output_coords[2])};
